backend/requirements.txt
alembic==1.13.1
celery==5.3.6
fastapi==0.110.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.28
passlib==1.7.4
bcrypt==4.1.2
requests==2.31.0
pytest==8.3.5
httpx==0.27.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.1.3
email-validator==2.1.1
python-dotenv==1.0.1
uvicorn==0.27.1 